migrating the api task example to be fully in its own namespace
